+++ CHANGELOG.md
@@ -23,6 +23,10 @@ For more information on SemVer, please visit http://semver.org/.
 ## Version 1.4.5
 ### 2013-11-28
 
+#### Enhancements
+
+* Added 16 new Glyphs Icons
+
 #### Improvements
 
 * Improved `wpdk_enqueue_script_page_template()` function